Delete tests fix

+++ candidate_review/src/main/webapp/controllers/testsControlController.js
@@ -15,20 +15,22 @@ angular.module('NESS-TCFA').controller('testsControlController',['$scope','$root
   );
   $scope.deleting = false;
   $scope.removeTest = function(test){
-    $scope.deleting = true;
-    $http.delete($rootScope.serverUrl+'admin/test/'+test.id+"?access_token="+$rootScope.settings.auth.access_token).then(
-      function(response){
-        $scope.deleting = false;
-        $scope.tests.splice($scope.tests.indexOf(test),1);
-      },
-      function(){
-        $scope.deleting = false;
-      }
-    );
+   if($scope.deleting == false){
+     $scope.deleting = true;
+     $http.delete($rootScope.serverUrl+'admin/test/'+test.id+"?access_token="+$rootScope.settings.auth.access_token).then(
+         function(response){
+           $scope.deleting = false;
+           $scope.tests.splice($scope.tests.indexOf(test),1);
+         },
+         function(){
+           $scope.deleting = false;
+         }
+     );
+   }
   };
   $scope.$on('$locationChangeStart', function( event ) {
      if ($scope.deleting == true) {
-       $window.location.href = '#/testsControl/';
+         event.preventDefault();
      }
    });
   $scope.activeTests = function(){
